# multisig-wallet-fe/swagger.yaml
openapi: 3.0.3
info:
  title: Welcome to Co-vault Protocol
  description: This is swagger UI for Co-vault protocol.
  version: 1.0.1
servers:
  - url: http://localhost:9015/api
...
paths:
...
  /airdrop/broadcasting:
    post:
      tags:
        - airdrop
      summary: broadcasting the transaction into mempool
      description: broadcasting the signed psbt into mempool
      operationId: broadcasting
      requestBody:
        description: after signing the psbt, broadcasting the transaction into mempool
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/broadcasting"
          application/xml:
            schema:
              $ref: "#/components/schemas/broadcasting"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/broadcasting"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/broadcasting200"
            application/xml:
              schema:
                $ref: "#/components/schemas/broadcasting200"
        "400":
          description: Invalid pubkeyList, minSignCount, assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/broadcasting400"
            application/xml:
              schema:
                $ref: "#/components/schemas/broadcasting400"
        "404":
          description: Vault not found
        "422":
          description: Validation exception
...
  /request/getAllRequestList:
    get:
      tags:
        - request
      summary: Returns all request List
      description: Returns all request List
      operationId: getAllRequestList
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getAllRequestList200"
            application/xml:
              schema:
                $ref: "#/components/schemas/getAllRequestList200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
        "422":
          description: Validation exception
  /request/getPsbtFromRequest:
    post:
      tags:
        - request
      summary: Returns specific request by id and pubkey
      description: Returns all request List
      operationId: getPsbtFromRequest
      requestBody:
        description: Get request by vault id and pubkey.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/getPsbtFromRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/getPsbtFromRequest400"
        "422":
          description: Validation exception
  /request/updateRequest:
    post:
      tags:
        - request
      summary: Sign request and broadcasting when threshold get reached
      description: Update reqeust
      operationId: updateRequest
      requestBody:
        description: vaultId of existed valut,  PublicKey of co-signers, minSignCount for Threshold, assets for DAO token, ordinalAddress of updater
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/updateRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/updateRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/updateRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/updateRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/updateRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/updateRequest400"
        "422":
          description: Validation exception
  /request/cancelUpdateForRequest:
    post:
      tags:
        - request
      summary: Cancel the request to update the psbt
      description: Cancel reqeust
      operationId: cancelUpdateForRequest
      requestBody:
        description: Cancle the request to update the psbt signing to let other can request to sign that psbt.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
          application/xml:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/cancelUpdateForRequest"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest200"
            application/xml:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest200"
        "400":
          description: There is no exist wallet with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest400"
            application/xml:
              schema:
                $ref: "#/components/schemas/cancelUpdateForRequest400"
        "422":
          description: Validation exception
...
components:
  schemas:
...
    Assets:
      required:
        - creatorAddress
        - initialPrice
        - runeAmount
        - runeName
        - runeSymbol
      type: object
      properties:
        creatorAddress:
          type: string
          example: tb1p2vdp0uvc68ng6j56mq4gdkd6jfwhj3qhq2zhp44zxnsgp903d9sqjqlq4r
        initialPrice:
          type: string
          example: 3000
        runeAmount:
          type: string
          example: 40000
        runeName:
          type: string
          example: EIGHTEEN.RUNE.TOKEN.TOKEN
        runeSymbol:
          type: string
          example: ✨
...
    getAllRequestList200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Fetch all request successfully.
        payload:
          type: array
          items:
            type: object
            properties:
              assets:
                $ref: "#/components/schemas/Assets"
              _id:
                type: string
                example: 668787fa55b8778f50981d34
              musigId:
                type: string
                example: 6686bf8b6683b6c56b119f43
              type:
                type: string
                example: VaultUpgrade
              transferAmount:
                type: string
                example: ALL
              destinationAddress:
                type: string
                example: tb1q4xs967j366zfzqzhazvpj555m2j43yvcvrhzhwgg6kgnrkh6av9qf73axx
              creator:
                type: string
                example: tb1p644m7l0fa4wv8x84w6wwhgh5alzpmnkvzxp95n776y6ek8348rwsamh86h
              cosigner:
                type: array
                items:
                  type: string
                example:
                  [
                    03abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a,
                    03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86,
                  ]
              signedCosigner:
                type: array
                items:
                  type: string
                example: []
              psbt:
                type: array
                items:
                  type: string
                example:
                  [
                    70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
                  ]
              threshold:
                type: number
                example: 2
              pending:
                type: string
                example: ""
              createdAt:
                type: string
                example: 2024-07-05T05:39:17.564Z
    getPsbtFromRequest:
      required:
        - id
        - pubKey
      type: object
      properties:
        id:
          type: string
          example: 6686bf8b6683b6c56b119f43
        pubkey:
          type: string
          example: 03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86
    getPsbtFromRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Fetch psbt successfully
        payload:
          type: string
          example: 70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000
    getPsbtFromRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Id or pubkey is missing.
        payload:
          type: object
          example: null
    updateRequest:
      required:
        - psbt
        - id
        - pubkey
      type: object
      properties:
        psbt:
          type: array
          items:
            example: 70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0220203abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a48304502210090abcf53964a34f18bb07b6b99cfb9993f7f03dcc6edd166001e67f254321e6102201a28da118400ef8979c55168cace37ec90094d46107c3fde715e34fb8a4d2f4a01010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000
          xml:
            wrapped: true
        id:
          type: string
          xml:
            wrapped: true
          example: 668791f3051db08c6aed8063
        pubkey:
          type: string
          example: 03abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a
    updateRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Request saved sucessfully
        payload:
          type: array
          items:
            example:
              [
                70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
                70736274ff0100fd310102000000038bf0969b932db9aae1b5532bcc64f8aeb55bab6a420014f66d94aece98cca5c20100000000ffffffff33e589a4f50c49becb7323dc6118c85bcc822f7ba70c0c11997c5b5e9203ce980200000000ffffffffec89f90656e564eff39f00b3276cd75afb3ad5fa2462ec8b8fd464e628ae61630000000000ffffffff049001000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a204e000000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a401f000000000000225120dd0db7020be2937c2554d68dc1be856d2483be077f00504b5790aba49fe29e5eca4c010000000000220020a9a05d7a51d684910057e898195294daa558919860ee2bb908d59131dafaeb0a000000000001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0220203abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a48304502210090abcf53964a34f18bb07b6b99cfb9993f7f03dcc6edd166001e67f254321e6102201a28da118400ef8979c55168cace37ec90094d46107c3fde715e34fb8a4d2f4a01010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012b2202000000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0001012ba086010000000000220020ae37dd3248538e8d802755f2c4008fd8613b6240ef890e73cfd682b3539f2aa0010569522103abb1b44f6526130e5b8b580b5f87f6a9668d6652ee336b282cb725be645aab6a2103481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a8621021fd26edce3324548565721728c72d9087471e127d3f4e093e78b04f4dad9aae853ae0000000000,
              ]
    updateRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Id, pubkey or psbt is missing.
        payload:
          type: object
          example: null
    cancelUpdateForRequest:
      required:
        - psbt
        - id
        - pubkey
      type: object
      properties:
        id:
          type: string
          xml:
            wrapped: true
          example: 668791f3051db08c6aed8063
        pubkey:
          type: string
          example: 03481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a86
    cancelUpdateForRequest200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Cancle request sucessfully
        payload:
          type: object
          example: null
    cancelUpdateForRequest400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: You didn't fetch this psbt
        payload:
          type: object
          example: null
...
    broadcasting:
      type: object
      properties:
        psbt:
          type: string
          example: 70736274ff0100f802000000020c96e9e1f970418e13a5b14dca0c3010688fe946e6221b14c47544dc3ba81d060100000000ffffffffa9ff140b27ecdc9f222d5a7b2ebe36f36d6b88467cf9e32c8ed3b79e563f6bbe0300000000ffffffff040000000000000000126a5d0f00dbfeab01f7110a020000fe9a04012202000000000000225120c4d10b54abbd8f9a6d536a40c2c098fac721d4667846ebb86a56f92f94a4d4042202000000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a94fa0f0000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a000000000001012b2202000000000000225120c4d10b54abbd8f9a6d536a40c2c098fac721d4667846ebb86a56f92f94a4d404011340f5bf53d9adc67ecc52d10a06112f8a795489a84d4348afea4473e194bdeb0512c1026b664e6b3047d28af4f89a40cebf67fd84477ade42426725e3d14c898a9e011720481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a860001012b54ee100000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a01172032362069574f88b2960e6c9c3491521b01ef32d913ec0f8ce6940eb89b7f7ec20000000000
        signedPsbt:
          type: string
          example: 70736274ff0100f802000000020c96e9e1f970418e13a5b14dca0c3010688fe946e6221b14c47544dc3ba81d060100000000ffffffffa9ff140b27ecdc9f222d5a7b2ebe36f36d6b88467cf9e32c8ed3b79e563f6bbe0300000000ffffffff040000000000000000126a5d0f00dbfeab01f7110a020000fe9a04012202000000000000225120c4d10b54abbd8f9a6d536a40c2c098fac721d4667846ebb86a56f92f94a4d4042202000000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a94fa0f0000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a000000000001012b2202000000000000225120c4d10b54abbd8f9a6d536a40c2c098fac721d4667846ebb86a56f92f94a4d404011340f5bf53d9adc67ecc52d10a06112f8a795489a84d4348afea4473e194bdeb0512c1026b664e6b3047d28af4f89a40cebf67fd84477ade42426725e3d14c898a9e011720481dd46ed4bfa803d6408a8294a678251a2869f8d24f0a0935899404b7f35a860001012b54ee100000000000225120e497e12e90a3a32c40568efccfe83e1c9da1c2e2269d62c1963564c888cda04a011340939b13aea169e4e3f03e41059564881f95df84acea51fbdb5068ac9963d75e005d594f7e286d5dc45cd21d5400497736ed26c4449e4a78f9ff71c49f3e6f218201172032362069574f88b2960e6c9c3491521b01ef32d913ec0f8ce6940eb89b7f7ec20000000000
        inputsCount:
          type: number
          example: 2
        walletType:
          type: string
          example: Unisat
        ordinalAddress:
          type: string
          example: tb1pujt7zt5s5w3jcszk3m7vl6p7rjw6rshzy6wk9svkx4jv3zxd5p9qzenv8v
        walletId:
          type: string
          example: 66793c473ea4fb1d998de3bb
    broadcasting200:
      type: object
      properties:
        success:
          type: boolean
          example: true
        msg:
          type: string
          example: e6b952f0b18832f77061b14cab890cccab795110c79badad11b6464f0dbfb034
    broadcasting400:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Get Error
        payload:
          type: object
          example: null
...